pok-08: refactor: Move pages to 'pages' directory
