+++ src/app/pages/favorites/favorites.page.scss
