+++ src/app/pages/favorites/favorites.page.spec.ts
