+++ src/app/pages/pokemon-details/pokemon-details.page.scss
