+++ src/app/pages/pokemon-details/pokemon-details.page.ts
@@ -23,14 +23,14 @@ import {
   IonFab,
   IonFabButton,
 } from '@ionic/angular/standalone';
-import { ToolbarComponent } from '../components/toolbar/toolbar.component';
+import { ToolbarComponent } from '../../components/toolbar/toolbar.component';
 import { ActivatedRoute, Router } from '@angular/router';
-import { SearchService } from '../services/search.service';
-import { Pokemon, PokemonEvolutionChain } from '../types/Pokemon';
+import { SearchService } from '../../services/search.service';
+import { Pokemon, PokemonEvolutionChain } from '../../types/Pokemon';
 
 import { addIcons } from 'ionicons';
 import { arrowBack, arrowForward, eye } from 'ionicons/icons';
-import { FavButtonComponent } from '../components/fav-button/fav-button.component';
+import { FavButtonComponent } from '../../components/fav-button/fav-button.component';
 
 @Component({
   selector: 'app-pokemon-details',
